QuestionTime/requirements.txt
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-registration==3.0
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.9.2
idna==2.8
oauthlib==3.0.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.2